Update params `CRISPResso_reference_info.txt` due to printing numpy arrays
Basically what is happening here is that the first amplicon's `include_idxs` and
`exclude_idxs` was always a numpy array, and the following amplicon's were
cloned off of it and were vanilla Python lists. All of the other arrays in this
file have commas separating them, but notice that these two columns didn't (but
they do now).